Add Oval as a supported shape
The OvalPanel in OvalPanel extends from the RectanglePanel
as the method of drawing a rectangle is the same as that of drawing an
oval. This allows us to reuse the mouse action listeners in the
RectanglePanel.

As the OvalPanel looks the best among the other Panels, I've also made
the OvalPanel as the default Panel when DrawingPanel opens up.

Also change the initial size of the rectangle to better show that it is
actually a rectangle. OvalPanel also benefits from this as it uses the
same rectangle.